move delaunay function
